service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-docx==0.8.11
weasyprint>=59.0
reportlab>=4.0.0
httpx>=0.25.0
